requirements.txt
pandas
psycopg2-binary
sqlalchemy 
python-dotenv